ai-rag-chatbot-project/requirements.txt
# ===== LangChain Core =====
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.30
langchain-experimental==0.3.4

# ===== Document Processing =====
pypdf==4.3.1
unstructured==0.14.6
langchain-text-splitters==0.3.9
PyMuPDF>=1.22.5


# ===== Embeddings & Vector Stores =====
sentence-transformers==2.7.0
huggingface_hub==0.25.1
faiss-cpu==1.8.0.post1
chromadb==0.5.5

# ===== LLM & AI Models =====
openai>=1.99.9,<2.0.0

# ===== Korean Language Processing =====
konlpy==0.6.0
langdetect==1.0.9

# ===== Utilities =====
python-dotenv==1.0.1
tiktoken==0.7.0
numpy==1.26.4
pandas==2.2.2
matplotlib==3.9.2
phonenumbers>=8.13.0

# ===== Database =====
psycopg2-binary==2.9.9

# ===== Additional Dependencies =====
transformers==4.44.2
torch==2.4.0
scikit-learn==1.5.1
scipy==1.13.1

# ===== Pydantic & Qdrant 호환 =====
pydantic==2.7.4
qdrant-client==1.12.0

# ===== 기타 충돌 해결 =====
fuzzytm==2.0.5
markdown-it-py==2.2.0
urllib3==2.0.7
fsspec==2023.10.0
pdfminer.six==20250327
websockets==11.0.3
cryptography==42.0.5
pillow==11.1.0
opentelemetry-semantic-conventions==0.54b1
importlib-metadata==6.8.0
packaging==23.2
protobuf==4.25.3
rich==13.9.2
tenacity==8.2.3
aiofiles==23.2.1
markupsafe==2.1.5